+++ lib/rails_cursor_pagination/paginator.rb
@@ -15,6 +15,9 @@ module RailsCursorPagination
     #
     # @param relation [ActiveRecord::Relation]
     #   Relation that will be paginated.
+    # @param limit [Integer, nil]
+    #   Number of records to return in pagination. Can be combined with either
+    #   `after` or `before` as an alternative to `first` or `last`.
     # @param first [Integer, nil]
     #   Number of records to return in a forward pagination. Can be combined
     #   with `after`.
@@ -35,14 +38,15 @@ module RailsCursorPagination
     # @param order [Symbol, nil]
     #   Ordering to apply, either `:asc` or `:desc`. Defaults to `:asc`.
     #
-    # @raise [RailsCursorPagination::Paginator::ParameterError]
+    # @raise [RailsCursorPagination::ParameterError]
     #   If any parameter is not valid
-    def initialize(relation, first: nil, after: nil, last: nil, before: nil,
-                   order_by: nil, order: nil)
+    def initialize(relation, limit: nil, first: nil, after: nil, last: nil,
+                   before: nil, order_by: nil, order: nil)
       order_by ||= :id
       order ||= :asc
 
-      ensure_valid_params!(relation, first, after, last, before, order)
+      ensure_valid_params_values!(relation, order, limit, first, last)
+      ensure_valid_params_combinations!(first, last, limit, before, after)
 
       @order_field = order_by
       @order_direction = order
@@ -54,6 +58,7 @@ module RailsCursorPagination
       @page_size =
         first ||
         last ||
+        limit ||
         RailsCursorPagination::Configuration.instance.default_page_size
 
       if Configuration.instance.max_page_size &&
@@ -80,26 +85,24 @@ module RailsCursorPagination
 
     private
 
-    # Ensure that the parameters of this service are valid. Otherwise raise
-    # a `RailsCursorPagination::Paginator::ParameterError`.
+    # Ensure that the parameters of this service have valid values, otherwise
+    # raise a `RailsCursorPagination::ParameterError`.
     #
     # @param relation [ActiveRecord::Relation]
     #   Relation that will be paginated.
-    # @param first [Integer, nil]
-    #   Optional, must be positive, cannot be combined with `last`
-    # @param after [String, nil]
-    #   Optional, cannot be combined with `before`
-    # @param last [Integer, nil]
-    #   Optional, must be positive, requires `before`, cannot be combined
-    #   with `first`
-    # @param before [String, nil]
-    #   Optional, cannot be combined with `after`
     # @param order [Symbol]
-    #   Optional, must be :asc or :desc
+    #   Must be :asc or :desc
+    # @param limit [Integer, nil]
+    #   Optional, must be positive
+    # @param first [Integer, nil]
+    #   Optional, must be positive
+    # @param last [Integer, nil]
+    #   Optional, must be positive
+    #   with `first` or `limit`
     #
-    # @raise [RailsCursorPagination::Paginator::ParameterError]
+    # @raise [RailsCursorPagination::ParameterError]
     #   If any parameter is not valid
-    def ensure_valid_params!(relation, first, after, last, before, order)
+    def ensure_valid_params_values!(relation, order, limit, first, last)
       unless relation.is_a?(ActiveRecord::Relation)
         raise ParameterError,
               'The first argument must be an ActiveRecord::Relation, but was ' \
@@ -109,20 +112,51 @@ module RailsCursorPagination
         raise ParameterError,
               "`order` must be either :asc or :desc, but was `#{order}`"
       end
+      if first.present? && first.negative?
+        raise ParameterError, "`first` cannot be negative, but was `#{first}`"
+      end
+      if last.present? && last.negative?
+        raise ParameterError, "`last` cannot be negative, but was `#{last}`"
+      end
+      if limit.present? && limit.negative?
+        raise ParameterError, "`limit` cannot be negative, but was `#{limit}`"
+      end
+
+      true
+    end
+
+    # Ensure that the parameters of this service are combined in a valid way.
+    # Otherwise raise a +RailsCursorPagination::ParameterError+.
+    #
+    # @param limit [Integer, nil]
+    #   Optional, cannot be combined with `last` or `first`
+    # @param first [Integer, nil]
+    #   Optional, cannot be combined with `last` or `limit`
+    # @param after [String, nil]
+    #   Optional, cannot be combined with `before`
+    # @param last [Integer, nil]
+    #   Optional, requires `before`, cannot be combined
+    #   with `first` or `limit`
+    # @param before [String, nil]
+    #   Optional, cannot be combined with `after`
+    #
+    # @raise [RailsCursorPagination::ParameterError]
+    #   If parameters are combined in an invalid way
+    def ensure_valid_params_combinations!(first, last, limit, before, after)
       if first.present? && last.present?
         raise ParameterError, '`first` cannot be combined with `last`'
+      end
+      if first.present? && limit.present?
+        raise ParameterError, '`limit` cannot be combined with `first`'
+      end
+      if last.present? && limit.present?
+        raise ParameterError, '`limit` cannot be combined with `last`'
       end
       if before.present? && after.present?
         raise ParameterError, '`before` cannot be combined with `after`'
       end
       if last.present? && before.blank?
         raise ParameterError, '`last` must be combined with `before`'
-      end
-      if first.present? && first.negative?
-        raise ParameterError, "`first` cannot be negative, but was `#{first}`"
-      end
-      if last.present? && last.negative?
-        raise ParameterError, "`last` cannot be negative, but was `#{last}`"
       end
 
       true